+++ injectable/example/lib/injector/injector.dart
@@ -1,3 +1,4 @@
+import 'package:awesome/awesome.module.dart';
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 import 'injector.config.dart';
@@ -5,7 +6,11 @@ import 'injector.config.dart';
 const platformMobile = Environment('platformMobile');
 const platformWeb = Environment('platformWeb');
 
-@InjectableInit()
+@InjectableInit(
+  externalPackageModulesAfter: [
+    ExternalModule(AwesomePackageModule),
+  ],
+)
 configInjector(
   GetIt getIt, {
   String? env,